Min. service rate a station must have to consider available cargo.
If a station doesn't create enough cargo to fill a vehicle, it is rather
seldom visited, so the service rate starts decreasing which reduces the
score of this station even further ... it gets quite imbalanced. This
change lets the user define a minimum service rate. If the service rate of
a station drops below that minimum, the supply is not longer taken into
account when calculating the scrore, just he service rate. The result is
that such a station gets a higher score again and visited which in turn
raises the service rate again.